# opt: reduce allocations in props.histogramIter