… pkg/sql/opt/memo/testdata/stats_quality/tpch/q06