This commit reduces allocations in props.histogramIter by avoiding allocations for bucket lower bounds. Instead, we use the upper bound of the previous bucket as an exclusive lower bound of the current bucket.

This significantly reduces the number allocations, especially when filtering a histogram by a span that includes many buckets. It reduces the number of bytes allocated, especially when the histogram bounds are larger values, such as strings larger than several bytes. The benefits can be observed in one of the optimizer micro-benchmarks:

```
name                                                                    old time/op    new time/op    delta
Phases/single-col-histogram-range/Simple/Parse-16                         12.9µs ± 5%    12.5µs ± 7%     ~     (p=0.421 n=5+5)
Phases/single-col-histogram-range/Simple/OptBuildNoNorm-16                95.5µs ± 4%    81.8µs ±10%  -14.35%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/OptBuildNorm-16                   148µs ± 3%     129µs ± 4%  -12.62%  (p=0.016 n=5+4)
Phases/single-col-histogram-range/Simple/Explore-16                        166µs ± 1%     144µs ± 2%  -13.48%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/ExecBuild-16                      176µs ± 3%     153µs ± 3%  -13.52%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Prepared/AssignPlaceholdersNoNorm-16    62.0µs ± 4%    47.9µs ± 6%  -22.75%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Prepared/AssignPlaceholdersNorm-16      61.8µs ± 0%    47.8µs ±12%  -22.64%  (p=0.016 n=4+5)
Phases/single-col-histogram-range/Prepared/Explore-16                     71.9µs ± 6%    61.5µs ± 1%  -14.47%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Prepared/ExecBuild-16                   80.6µs ± 2%    67.8µs ± 5%  -15.78%  (p=0.008 n=5+5)

name                                                                    old alloc/op   new alloc/op   delta
Phases/single-col-histogram-range/Simple/Parse-16                         3.25kB ± 0%    3.25kB ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Simple/OptBuildNoNorm-16                33.5kB ± 0%    17.9kB ± 0%  -46.53%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/OptBuildNorm-16                  57.4kB ± 0%    26.2kB ± 0%  -54.38%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-16                       60.2kB ± 0%    29.0kB ± 0%  -51.82%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/ExecBuild-16                     61.7kB ± 0%    30.5kB ± 0%  -50.57%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Prepared/AssignPlaceholdersNoNorm-16    26.0kB ± 0%    10.5kB ± 0%  -59.83%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Prepared/AssignPlaceholdersNorm-16      26.0kB ± 0%    10.5kB ± 0%  -59.83%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Prepared/Explore-16                     28.3kB ± 0%    12.7kB ± 0%  -55.07%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Prepared/ExecBuild-16                   29.8kB ± 0%    14.2kB ± 0%  -52.32%  (p=0.008 n=5+5)

name                                                                    old allocs/op  new allocs/op  delta
Phases/single-col-histogram-range/Simple/Parse-16                           20.0 ± 0%      20.0 ± 0%     ~     (all equal)
Phases/single-col-histogram-range/Simple/OptBuildNoNorm-16                   608 ± 0%       282 ± 0%  -53.62%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/OptBuildNorm-16                   1.15k ± 0%     0.50k ± 0%  -56.55%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/Explore-16                        1.16k ± 0%     0.51k ± 0%  -56.06%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Simple/ExecBuild-16                      1.17k ± 0%     0.52k ± 0%  -55.63%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Prepared/AssignPlaceholdersNoNorm-16       564 ± 0%       238 ± 0%  -57.80%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Prepared/AssignPlaceholdersNorm-16         564 ± 0%       238 ± 0%  -57.80%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Prepared/Explore-16                        573 ± 0%       247 ± 0%  -56.89%  (p=0.008 n=5+5)
Phases/single-col-histogram-range/Prepared/ExecBuild-16                      582 ± 0%       256 ± 0%  -56.01%  (p=0.008 n=5+5)
```

Fixes #89982

Epic: None

Release note: None