@mgartner 2022-12-14
A better solution might be to try to allocate the lower bound lazily, and only do so when the bucket needs to be filtered. Buckets that are not filtered at all shouldn't need to allocate the lower bound.