@mgartner 2022-12-15
Closing in favor of the simpler solution in #93743.